Elimination of the recall of an AR
It makes absolutely no sense to have this in the constitution. If a person gets voted into office by the public they shouldn't be able to get recalled by the public. This makes it so small parties or the opposition can have many of their ARs taken away by a more popular party.